G2P infinite loop on non-UTF-8 files
It appears our G2P program in Java collapses if it fins any character that is
not a lower case letter. If it finds a simple space it enter into a infinite
loop wthout any king of warning. This also happens when the encoding of the
input file (list of words) is not UTF-8. Since a commong encoding type for PT_BR
chars is ISO-8859-1, it also represents an issue for the G2P software.